+++ docker/PhotoMigrator.sh
@@ -6,20 +6,17 @@ RELEASE_TAG="latest-stable"
 # Set a default TZ (e.g. UTC) in case docker.conf does not provide one
 TZ="UTC"
 
+# Linux image repo
+IMAGE_REPO="jaimetur/photomigrator-linux"
+
 # Load from docker.conf if exists and
-# Manually parse docker.conf to support inline and full-line comments
+# manually parse docker.conf to support inline and full-line comments
 if [ -f "docker.conf" ]; then
   while IFS= read -r line; do
-    # Remove leading/trailing whitespace
     line=$(echo "$line" | sed 's/^[[:space:]]*//;s/[[:space:]]*$//')
-
-    # Skip empty lines or full-line comments
     [[ -z "$line" || "$line" == \#* ]] && continue
-
-    # Remove inline comment after value
     line=$(echo "$line" | cut -d '#' -f1 | sed 's/[[:space:]]*$//')
 
-    # Parse key=value pairs
     key=$(echo "$line" | cut -d '=' -f1 | sed 's/[[:space:]]//g')
     value=$(echo "$line" | cut -d '=' -f2- | sed 's/^[[:space:]]*//;s/[[:space:]]*$//')
 
@@ -31,14 +28,14 @@ if [ -f "docker.conf" ]; then
   done < docker.conf
 fi
 
-# Get current directory
 CURRENT_DIR="$(pwd)"
 
-echo "🐳 Pulling Docker image: jaimetur/photomigrator:${RELEASE_TAG}"
-docker pull "jaimetur/photomigrator:${RELEASE_TAG}"
+echo "Pulling Docker image: ${IMAGE_REPO}:${RELEASE_TAG}"
+docker pull "${IMAGE_REPO}:${RELEASE_TAG}"
 
-echo "🚀 Launching container with TAG='${RELEASE_TAG}' and TZ='${TZ}'..."
+echo "Launching container with TAG='${RELEASE_TAG}' and TZ='${TZ}'..."
 docker run -it --rm \
   -v "$CURRENT_DIR":/docker \
   -e TZ="${TZ}" \
-  "jaimetur/photomigrator:${RELEASE_TAG}" "$@"
+  "${IMAGE_REPO}:${RELEASE_TAG}" "$@"
+