+++ src/Core/GlobalVariables.py
@@ -10,7 +10,7 @@ from colorama import Fore
 #---------------------------------------
 COPYRIGHT_TEXT                  = "(c) 2024-2026 - Jaime Tur (@jaimetur)"
 TOOL_NAME                       = "PhotoMigrator"
-TOOL_VERSION_WITHOUT_V          = "3.7.1"
+TOOL_VERSION_WITHOUT_V          = "3.8.0"
 TOOL_VERSION                    = f"v{TOOL_VERSION_WITHOUT_V}"
 TOOL_DATE                       = "2026-03-16"
 TOOL_NAME_VERSION               = f"{TOOL_NAME} {TOOL_VERSION}"