+++ src/web_interface/__init__.py
@@ -1,0 +1,2 @@
+"""PhotoMigrator web interface package."""
+